sandbox/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
base58==2.1.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.3
dnspython==2.6.1
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.109.2
gunicorn==21.2.0
h11==0.14.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
pycparser==2.21
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
pyOpenSSL==24.0.0
python-dotenv==1.0.1
requests==2.31.0
secp256k1==0.14.0
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.27.0.post1
joserfc==0.9.0
py-multibase==1.0.3